- Fail less noisily if "for some reason" input files disappear from the workspace. - Don't NPE on Step/Resume when no debugger is running.
git-svn-id: https://repos.hats-project.eu:444/svn/hats/Tools/ABS/trunk@13663 97f52227-7fa9-4375-a72d-c1c733b1f48f

+++ eclipse-plugin/src/eu/hatsproject/absplugin/builder/AbsNature.java
@@ -266,8 +266,9 @@ public class AbsNature implements IProjectNature {
 	 * @throws CoreException @{@link IResource#deleteMarkers(String, boolean, int)} 
 	 */
 	public void parseABSFile(IResource resource, boolean withincomplete, IProgressMonitor monitor) throws CoreException {
-		if (isABSFile(resource)) {
+		if (resource.exists() && isABSFile(resource)) {
 			IFile file = (IFile) resource;
+			assert file.exists();
 			file.deleteMarkers(MARKER_TYPE, true, IResource.DEPTH_ZERO);
 			try {
 			   if (!file.isSynchronized(IResource.DEPTH_ZERO)) {
